--- dribas.sergey/P3/checkEnter.hpp
@@ -1,9 +1,0 @@
-#ifndef CHECKENTER_HPP
-#define CHECKENTER_HPP
-
-namespace dribas
-{
-  bool checkEnter(char* const* const argv, int argc, long int& secondArg);
-}
-
-#endif
